REFACTOR renamed client classes
